Move version bumper method to library class
